adds simulation service with monthly payment calculation and test class

+++ src/main/kotlin/com/project/loan_simulator/dto/SimulationRequest.kt
@@ -4,6 +4,7 @@ import com.project.loan_simulator.validation.MinimumAge
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.validation.constraints.Positive
 import org.jetbrains.annotations.NotNull
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Schema(description = "Model for requesting a loan simulation")
@@ -11,11 +12,11 @@ data class SimulationRequest(
     @field:Schema(
         description = "Total value requested to loan",
         example = "10000",
-        type = "int",
+        type = "BigDecimal",
     )
     @field:NotNull
     @field:Positive
-    val totalValue: Int,
+    val totalValue: BigDecimal,
     @field:Schema(
         description = "Date of birth of the client, must be sent as a string following the format YYYY-MM-DD",
         example = "1989-05-27",
@@ -27,7 +28,7 @@ data class SimulationRequest(
     @field:Schema(
         description = "The length of the monthly payments to be made",
         example = "36",
-        type = "int",
+        type = "Int",
     )
     @field:NotNull
     @field:Positive
